FreeType.LowLevel.GlyphMetrics

Documentation

data GlyphMetrics a

Wrapper for FT_Glyph_Metrics.

Constructors

GlyphMetrics

Fields

width :: a

The glyph's width.

height :: a

The glyph's height.

horiBearingX :: a

Left side bearing for horizontal layout.

horiBearingY :: a

Top side bearing for horizontal layout.

horiAdvance :: a

Advance width for horizontal layout.

vertBearingX :: a

Left side bearing for vertical layout.

vertBearingY :: a

Top side bearing for vertical layout. Larger positive values mean further below the vertical glyph origin.

vertAdvance :: a

Advance height for vertical layout. Positive values mean the glyph has a positive advance downward.